# back/modules/generated/public/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "email" : "game@frol.be"
    },
    "description" : "online game platform",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "Game",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://game.frol.be/api"
  }, {
    "url" : "http://game.frol.be/api"
  } ],
...
  "paths" : {
...
    "/socialLogin" : {
      "post" : {
        "operationId" : "socialLogin",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserAssociation"
              }
            }
          },
          "description" : "the association infos",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : { },
            "description" : "ok"
          }
        },
        "summary" : "social login",
        "tags" : [ "auth" ],
        "x-codegen-request-body-name" : "userAssociation"
      }
    },
...
    "/user/link" : {
      "post" : {
        "operationId" : "linkToExternalAccount",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserAssociation"
              }
            }
          },
          "description" : "the association infos",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "link to external account",
        "tags" : [ "auth" ],
        "x-codegen-request-body-name" : "userAssociation"
      }
    },
...
    "/game/{uuid}" : {
      "get" : {
        "operationId" : "getGame",
        "parameters" : [ {
          "in" : "path",
          "name" : "uuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Game"
                }
              }
            },
            "description" : "ok"
          }
        },
        "summary" : "get a games",
        "tags" : [ "game" ]
      }
    },
...
    "/game/LostInTranslation/{uuid}" : {
      "get" : {
        "operationId" : "getGame",
        "parameters" : [ {
          "in" : "path",
          "name" : "uuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LostInTranslationGame"
                }
              }
            },
            "description" : "ok"
          }
        },
        "summary" : "Get an LostInTranslation game",
        "tags" : [ "lostInTranslation" ]
      }
    },
    "/game/LostInTranslation/{uuid}/story/{storyId}/text" : {
      "post" : {
        "operationId" : "addTextRound",
        "parameters" : [ {
          "in" : "path",
          "name" : "uuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "storyId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "description" : "the text",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LostInTranslationGame"
                }
              }
            },
            "description" : "ok"
          }
        },
        "summary" : "add text round in LostInTranslation game",
        "tags" : [ "lostInTranslation" ],
        "x-codegen-request-body-name" : "text"
      }
    },
    "/game/LostInTranslation/{uuid}/story/{storyId}/drawing" : {
      "post" : {
        "operationId" : "addDrawingRound",
        "parameters" : [ {
          "in" : "path",
          "name" : "uuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "storyId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "file" : {
                    "format" : "binary",
                    "type" : "string"
                  }
                },
                "required" : [ "file" ]
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LostInTranslationGame"
                }
              }
            },
            "description" : "ok"
          }
        },
        "summary" : "add drawing round in lostInTranslation game",
        "tags" : [ "lostInTranslation" ]
      }
    },
    "/game/DefineItLy/{uuid}" : {
      "get" : {
        "operationId" : "getGame",
        "parameters" : [ {
          "in" : "path",
          "name" : "uuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DefineItLyGame"
                }
              }
            },
            "description" : "ok"
          }
        },
        "summary" : "Get a DefineItLy game",
        "tags" : [ "defineitly" ]
      }
    },
    "/game/DefineItLy/{uuid}/question" : {
      "post" : {
        "operationId" : "addQuestion",
        "parameters" : [ {
          "in" : "path",
          "name" : "uuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StringWrapper"
              }
            }
          },
          "description" : "the question",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DefineItLyGame"
                }
              }
            },
            "description" : "ok"
          }
        },
        "summary" : "Add a definitely question",
        "tags" : [ "defineitly" ],
        "x-codegen-request-body-name" : "text"
      }
    },
    "/game/DefineItLy/{uuid}/question/{questionUuid}" : {
      "post" : {
        "operationId" : "addResponse",
        "parameters" : [ {
          "in" : "path",
          "name" : "uuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "questionUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StringWrapper"
              }
            }
          },
          "description" : "the response",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DefineItLyGame"
                }
              }
            },
            "description" : "ok"
          }
        },
        "summary" : "Add a definitely response to a question",
        "tags" : [ "defineitly" ],
        "x-codegen-request-body-name" : "text"
      }
    },
    "/game/DefineItLy/{uuid}/question/{questionUuid}/choice" : {
      "post" : {
        "operationId" : "choseResponse",
        "parameters" : [ {
          "in" : "path",
          "name" : "uuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "questionUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StringWrapper"
              }
            }
          },
          "description" : "the response uuid",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DefineItLyGame"
                }
              }
            },
            "description" : "ok"
          }
        },
        "summary" : "Chose a response to a question",
        "tags" : [ "defineitly" ],
        "x-codegen-request-body-name" : "responseUuid"
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "User" : {
        "example" : {
          "canAdministrageGame" : true,
          "name" : "name",
          "linkedAccounts" : [ {
            "linkType" : "Facebook",
            "accessToken" : "accessToken"
          }, {
            "linkType" : "Facebook",
            "accessToken" : "accessToken"
          } ],
          "uuid" : "uuid"
        },
        "properties" : {
          "uuid" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "canAdministrageGame" : {
            "type" : "boolean"
          },
          "linkedAccounts" : {
            "items" : {
              "$ref" : "#/components/schemas/UserAssociation"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "UserAssociation" : {
        "example" : {
          "linkType" : "Facebook",
          "accessToken" : "accessToken"
        },
        "properties" : {
          "linkType" : {
            "enum" : [ "Facebook" ],
            "type" : "string"
          },
          "accessToken" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "GameDescription" : {
        "example" : {
          "gameType" : "lostInTranslation",
          "startTime" : "2000-01-23T04:56:07.000+00:00",
          "id" : 0,
          "uuid" : "uuid",
          "status" : "toStart"
        },
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "uuid" : {
            "type" : "string"
          },
          "startTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "toStart", "inPlay", "finished" ],
            "type" : "string"
          },
          "gameType" : {
            "enum" : [ "lostInTranslation", "define-it-ly" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "Game" : {
        "example" : {
          "description" : {
            "gameType" : "lostInTranslation",
            "startTime" : "2000-01-23T04:56:07.000+00:00",
            "id" : 0,
            "uuid" : "uuid",
            "status" : "toStart"
          },
          "users" : [ {
            "canAdministrageGame" : true,
            "name" : "name",
            "linkedAccounts" : [ {
              "linkType" : "Facebook",
              "accessToken" : "accessToken"
            }, {
              "linkType" : "Facebook",
              "accessToken" : "accessToken"
            } ],
            "uuid" : "uuid"
          }, {
            "canAdministrageGame" : true,
            "name" : "name",
            "linkedAccounts" : [ {
              "linkType" : "Facebook",
              "accessToken" : "accessToken"
            }, {
              "linkType" : "Facebook",
              "accessToken" : "accessToken"
            } ],
            "uuid" : "uuid"
          } ]
        },
        "properties" : {
          "description" : {
            "$ref" : "#/components/schemas/GameDescription"
          },
          "users" : {
            "items" : {
              "$ref" : "#/components/schemas/User"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "LostInTranslationGame" : {
        "example" : {
          "game" : {
            "description" : {
              "gameType" : "lostInTranslation",
              "startTime" : "2000-01-23T04:56:07.000+00:00",
              "id" : 0,
              "uuid" : "uuid",
              "status" : "toStart"
            },
            "users" : [ {
              "canAdministrageGame" : true,
              "name" : "name",
              "linkedAccounts" : [ {
                "linkType" : "Facebook",
                "accessToken" : "accessToken"
              }, {
                "linkType" : "Facebook",
                "accessToken" : "accessToken"
              } ],
              "uuid" : "uuid"
            }, {
              "canAdministrageGame" : true,
              "name" : "name",
              "linkedAccounts" : [ {
                "linkType" : "Facebook",
                "accessToken" : "accessToken"
              }, {
                "linkType" : "Facebook",
                "accessToken" : "accessToken"
              } ],
              "uuid" : "uuid"
            } ]
          },
          "stories" : [ {
            "storyId" : "storyId",
            "originalUser" : {
              "canAdministrageGame" : true,
              "name" : "name",
              "linkedAccounts" : [ {
                "linkType" : "Facebook",
                "accessToken" : "accessToken"
              }, {
                "linkType" : "Facebook",
                "accessToken" : "accessToken"
              } ],
              "uuid" : "uuid"
            },
            "rounds" : [ {
              "storyId" : "storyId",
              "drawing" : "drawing",
              "roundUser" : {
                "canAdministrageGame" : true,
                "name" : "name",
                "linkedAccounts" : [ {
                  "linkType" : "Facebook",
                  "accessToken" : "accessToken"
                }, {
                  "linkType" : "Facebook",
                  "accessToken" : "accessToken"
                } ],
                "uuid" : "uuid"
              },
              "submissionDate" : "2000-01-23T04:56:07.000+00:00",
              "roundType" : "text",
              "text" : "text"
            }, {
              "storyId" : "storyId",
              "drawing" : "drawing",
              "roundUser" : {
                "canAdministrageGame" : true,
                "name" : "name",
                "linkedAccounts" : [ {
                  "linkType" : "Facebook",
                  "accessToken" : "accessToken"
                }, {
                  "linkType" : "Facebook",
                  "accessToken" : "accessToken"
                } ],
                "uuid" : "uuid"
              },
              "submissionDate" : "2000-01-23T04:56:07.000+00:00",
              "roundType" : "text",
              "text" : "text"
            } ]
          }, {
            "storyId" : "storyId",
            "originalUser" : {
              "canAdministrageGame" : true,
              "name" : "name",
              "linkedAccounts" : [ {
                "linkType" : "Facebook",
                "accessToken" : "accessToken"
              }, {
                "linkType" : "Facebook",
                "accessToken" : "accessToken"
              } ],
              "uuid" : "uuid"
            },
            "rounds" : [ {
              "storyId" : "storyId",
              "drawing" : "drawing",
              "roundUser" : {
                "canAdministrageGame" : true,
                "name" : "name",
                "linkedAccounts" : [ {
                  "linkType" : "Facebook",
                  "accessToken" : "accessToken"
                }, {
                  "linkType" : "Facebook",
                  "accessToken" : "accessToken"
                } ],
                "uuid" : "uuid"
              },
              "submissionDate" : "2000-01-23T04:56:07.000+00:00",
              "roundType" : "text",
              "text" : "text"
            }, {
              "storyId" : "storyId",
              "drawing" : "drawing",
              "roundUser" : {
                "canAdministrageGame" : true,
                "name" : "name",
                "linkedAccounts" : [ {
                  "linkType" : "Facebook",
                  "accessToken" : "accessToken"
                }, {
                  "linkType" : "Facebook",
                  "accessToken" : "accessToken"
                } ],
                "uuid" : "uuid"
              },
              "submissionDate" : "2000-01-23T04:56:07.000+00:00",
              "roundType" : "text",
              "text" : "text"
            } ]
          } ]
        },
        "properties" : {
          "game" : {
            "$ref" : "#/components/schemas/Game"
          },
          "stories" : {
            "items" : {
              "$ref" : "#/components/schemas/LostInTranslationStory"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "LostInTranslationStory" : {
        "example" : {
          "storyId" : "storyId",
          "originalUser" : {
            "canAdministrageGame" : true,
            "name" : "name",
            "linkedAccounts" : [ {
              "linkType" : "Facebook",
              "accessToken" : "accessToken"
            }, {
              "linkType" : "Facebook",
              "accessToken" : "accessToken"
            } ],
            "uuid" : "uuid"
          },
          "rounds" : [ {
            "storyId" : "storyId",
            "drawing" : "drawing",
            "roundUser" : {
              "canAdministrageGame" : true,
              "name" : "name",
              "linkedAccounts" : [ {
                "linkType" : "Facebook",
                "accessToken" : "accessToken"
              }, {
                "linkType" : "Facebook",
                "accessToken" : "accessToken"
              } ],
              "uuid" : "uuid"
            },
            "submissionDate" : "2000-01-23T04:56:07.000+00:00",
            "roundType" : "text",
            "text" : "text"
          }, {
            "storyId" : "storyId",
            "drawing" : "drawing",
            "roundUser" : {
              "canAdministrageGame" : true,
              "name" : "name",
              "linkedAccounts" : [ {
                "linkType" : "Facebook",
                "accessToken" : "accessToken"
              }, {
                "linkType" : "Facebook",
                "accessToken" : "accessToken"
              } ],
              "uuid" : "uuid"
            },
            "submissionDate" : "2000-01-23T04:56:07.000+00:00",
            "roundType" : "text",
            "text" : "text"
          } ]
        },
        "properties" : {
          "originalUser" : {
            "$ref" : "#/components/schemas/User"
          },
          "storyId" : {
            "type" : "string"
          },
          "rounds" : {
            "items" : {
              "$ref" : "#/components/schemas/LostInTranslationRound"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "LostInTranslationRound" : {
        "example" : {
          "storyId" : "storyId",
          "drawing" : "drawing",
          "roundUser" : {
            "canAdministrageGame" : true,
            "name" : "name",
            "linkedAccounts" : [ {
              "linkType" : "Facebook",
              "accessToken" : "accessToken"
            }, {
              "linkType" : "Facebook",
              "accessToken" : "accessToken"
            } ],
            "uuid" : "uuid"
          },
          "submissionDate" : "2000-01-23T04:56:07.000+00:00",
          "roundType" : "text",
          "text" : "text"
        },
        "properties" : {
          "roundUser" : {
            "$ref" : "#/components/schemas/User"
          },
          "storyId" : {
            "type" : "string"
          },
          "submissionDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "roundType" : {
            "enum" : [ "text", "drawing" ],
            "type" : "string"
          },
          "text" : {
            "type" : "string"
          },
          "drawing" : {
            "description" : "id of the drawing file, use file api to recover it",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "StringWrapper" : {
        "example" : {
          "value" : "value"
        },
        "properties" : {
          "value" : {
            "type" : "string"
          }
        },
        "required" : [ "value" ],
        "type" : "object"
      }
    }
  }
}